==> Revisao_Literatura/google_scholar/1 - Crawler Google Scholar.ipynb <==
import time
import pandas as pd
from tqdm.notebook import tqdm
from selenium import webdriver
import numpy as np 
from pathlib import Path
from unidecode import unidecode
import os 
import undetected_chromedriver as uc

path = 'data'

root_folder = Path(path)
root_folder.mkdir(parents=True, exist_ok=True)

df_searchterms = pd.read_csv('../searchterms.csv')
df_seq = pd.DataFrame(np.arange(0,10000,10), columns = ['start'])

df_searchterms['k']=1
df_seq['k']=1

df_searchterms = df_searchterms.merge(df_seq, on ='k').drop(columns = 'k')
df_searchterms['searchterm'] = df_searchterms['searchterm'].map(lambda x: unidecode(x)).str.lower().str.strip()
df_searchterms['folder'] = df_searchterms['searchterm'].str.replace(' ','_',regex = False)
df_searchterms['url'] = "https://scholar.google.com.br/scholar?start="+ df_searchterms['start'].astype('str')+"&q="+df_searchterms['searchterm'].str.replace(' ','+',regex = False)+"&hl=pt-BR&as_sdt=0,5"
df_searchterms['file'] = (
    df_searchterms['url']
    .str.replace(r'[^a-zA-Z0-9_]', '_', regex=True)  # substitui caracteres especiais por _
    .str.replace(r'_+', '_', regex=True)             # múltiplos _ por 1 _
    .str.strip('_')                                  # remove _ no começo/fim
    + '.html'
)

#LIMITE PARA TESTES
df_searchterms = df_searchterms[df_searchterms['start'] < 300].copy()

df_searchterms.to_csv(path+'/google_scholar_search_base.csv', sep = ';', index = False)


# Inicializa o navegador
#driver = webdriver.Chrome()
driver = uc.Chrome(version_main=132)
url = f"https://scholar.google.com.br/scholar?start=0&q=visualization+machine+learning&hl=pt-BR&as_sdt=0,5"
driver.get(url)



driver.get(url)

records_search = df_searchterms.to_dict(orient = 'records')
records_search_filter = []
for record in records_search:
    local_folder = caminho_completo = root_folder /'raw_data' / record['folder']
    local_folder.mkdir(parents=True, exist_ok=True)

    local_file = local_folder / record['file']
    if local_file.exists() and local_file.is_file():  
        continue

    records_search_filter.append((record, local_file))

for record, local_file in tqdm(records_search_filter):

    driver.get(record['url'])
    time.sleep(2)
    alpha, beta = 3,6
    val = np.random.beta(alpha, beta)
    wait_time = 1 + val * (3 - 1)
    time.sleep(wait_time) 

    if 'Confirme que você é humano' in driver.page_source or\
                'Verificando se você é humano' in driver.page_source or\
                'There was a problem providing the content you requested' in driver.page_source or\
                'but we are experiencing unusual' in driver.page_source or\
                'não é um robô' in driver.page_source or\
                'Nossos sistemas detectaram tráfego incomum' in driver.page_source or\
                'we require CAPTCHA on specific webpages' in driver.page_source:
                break

    with open(local_file, "w", encoding="utf-8") as f:
        f.write(driver.page_source)

    

#for i in range(len(records_search_filter)):
#    record, local_file = records_search_filter[i]
#    local_file.unlink()
